Use `git clone` to clone the repo
This allows us to use full commit ids, short hashes and tags for the
components, and also allows us to fetch the correct image by calculating
the image version from the commit itself (in case its not a tag).

This should allow us to use `gh release` with short commit ids (which it
doesnt support).

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>