Fix release to properly release images
Release was trying to release images without pulling them.
Release now depends on images which will always try to pull the
necessary images.
As a consequence, we can now always run E2E with master branch in case
we're in a partial stage and some of the components aren't known yet.

Also fixed subctl building which was broken due to recent changes in
submariner-operator.
It should now work and be more robust, as we're essentially running the
whole build "dapper in dapper".

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>