Fix early release of Shipyard images
During releases, we try to release Shipyard images early so that the
validation jobs spawned by the release don't fail because of missing
images. This relied on the fact that "make release-images" would
blindly release any locally-available image matching the name of the
image(s) it was given (including in RELEASE_ARGS). This broke with the
introduction of multi-arch image support: each project now filters its
images based on whether they're multi-arch or not, using the
MULTIARCH_IMAGES and IMAGES variables; but the actual release here
runs in the context of the "releases" project, not the project
actually producing the images, so nothing matches and nothing gets
released.

Move the image release inside the Shipyard project so that the correct
architecture determination can be made. This ends up releasing nettest
too, which we don't need, but this doesn't cause any harm.

Signed-off-by: Stephen Kitt <skitt@redhat.com>