Branch out `releases` on `rc0`
Since we're creating stable branches on `rc0` we should create the
branch on `releases` as well, so that all stable releases happen on the
branch.

This step must happen on the CI since:
1. It creates a branch owned by the bot, not an individual contributor.
2. At this point, we're creating branches for all projects, which will
   trigger image builds for these branches.
   These images will then be available for any further release on that
   stable branch.

This also makes sure the releases on `releases` happen on the stable
branch.
When releasing a stable release, it must happen on the stable branch
itself (should it exist).

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>