Fix `determine_target_release` merge detection
Sometimes using "Update branch" in GitHub causes a "double" merge commit
event, which trips up the release detection logic.

Example:
```
commit ba36cd7a36b0bfd05122759b33f5509b47bcc8d8 (HEAD, pull/381/merge)
Merge: f4f375f 302ed6d
Author: Mike Kolesnik <mkolesni@redhat.com>
Date:   Tue Jun 21 08:42:09 2022 +0300

    Merge 302ed6d0d436261851023deb8c14c5d7271829b6 into \
f4f375f772506de5c8f10423c6417aa4a6c6b74a

commit 302ed6d0d436261851023deb8c14c5d7271829b6 (origin/releasing-0.13.0-rc0)
Merge: 2a94578 f4f375f
Author: Mike Kolesnik <mkolesni@redhat.com>
Date:   Tue Jun 21 08:42:08 2022 +0300

    Merge branch 'release-0.13' into releasing-0.13.0-rc0
```

Fixed the logic to account for the possibility that the merged commit is
a merge commit by itself.
In this case, the 1st parent of the 2nd parent of the original commit,
is the tip of the PR we're updating.

In the example, the 2nd parent is `302ed6d` which in itself is a merge
commit where the 1st parent `2a94578` is the PR branch
(releasing-0.13.0-rc0).

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>