Adjust to latest Shipyard changes
Shipyard changed the way it sets images, to use `PRELOAD_IMAGES`.
As such, adjust `releases` to set the variable when necessary, and
remove out custom preloading script.

This is a stop gap solution for releasing, a proper solution which
pre-loads just the necessary images will need to be worked out.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>