Run `go mod` updates when testing
Right now the dependency updates are being skipped since the versions
aren't really being release, but this causes us to miss testing this
part of the release process, which could have some stop stopping errors.

This fix will run the updates on the latest stable branch version when
we're dry-running, allowing us to validate this part as well.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>