Bump all dependencies in one go
Instead of trickling down dependency bumps, this applies all
dependency upgrades in a single PR per project. This doesn't change
anything for shipyard; then:
* once shipyard is released, only admiral is upgraded
* once admiral is released, cloud-prepare, lighthouse and submariner
  are upgraded (with the new releases of both shipyard and admiral)
* once cloud-prepare, lighthouse, and submariner are released,
  submariner-operator is upgraded
* once submariner-operator is released, subctl is upgraded

This reduces the overall number of PRs needed for a release, and
ensures that each project gets a consistent set of upgrades (avoiding
breakage related e.g. to API changes in one project requiring
simultaneous upgrades of all affected dependencies in the operator or
subctl).

Signed-off-by: Stephen Kitt <skitt@redhat.com>